backend/entity_identity/requirements.txt
openai
fastapi
uvicorn
mysql-connector-python
google-generativeai